GoogleTTS/requirements.txt
google-cloud-texttospeech==2.16.3
python-dotenv==1.0.0